US7E1 - add a new genre
In the GenreController, I have implemented a method that uses MediatR to send the corresponding CreateGenreCommand to the CreateGenreCommandHandler. This handler then interacts with the GenresRepository to perform the necessary operations. The implementation is straightforward, leveraging the existing boilerplate code created for a similar user story, ensuring consistency and adherence to clean architecture principles.

Now the Genre model is not anemic, consequently, the GenreSeeder was
refactored, and a new migration created.