US2E1.1 - cache for total number of games header
I created a new project called "Infrastructure" for the date and time service. Maybe this is overkill; I had previously separated the persistence into its own project, so I didn't want to use that project for placing the date and time service. I just introduced basic logic for checking if the value was stored before sending the query in the total games header middleware, which was previously implemented for getting the game count.