US3E2 - add a new publisher endpoint
The web application, domain, and persistence layer boilerplate were implemented, but this time I moved the validations to the application layer using the FluentValidation library. Specifically, I created the CreatePublisherCommandBehavior for validating the CreatePublisherCommand instead of the CreatePublisherRequest. I'm not sure about this change because I preferred failing fast when it was possible in the controller, but I think validating the Command is acceptable.