US7E2 - update Publisher
I created the UpdatePublisherCommandBehavior to validate the UpdatePublisherCommand instead of the UpdatePublisherRequest. This means that the validations are made in the application layer. The implementation with the basic validations was relatively straightforward. There might be some duplicated validations, and introducing a value object like URL or something similar could be a good idea for future improvement