US1E6 - Get Shippers endpoint
I implemented the required feature using the MongoDb.Driver NuGet package. For this, I created a Shipper model in both the Persistence and Domain projects, each with a defined concern. The Persistence project is responsible for receiving data from the MongoDB database, while the Domain project manages the domain logic. In this case, the domain model is anemic, but I followed this approach to maintain the separation of concerns.

I also implemented the necessary Controller, Request, Request Handler, Repository, and other components required for this task.