US2E6 - Get orders history
I created a read-only repository for the MongoDB database and implemented a custom serializer for the dates. Since I haven't implemented authorization and MongoDB doesn't use GUIDs for the order IDs, I used a dummy GUID for the customer ID and another dummy GUID for the order ID. After that, I combined the orders from the SQL Server database and the MongoDB database as the result of the API call.