US5E6 - Implement unit-in-stock count sync for two databases.
Implemented a repository to read the current games from the MongoDB Products collection. If a game exists in the SQL Server database, the unit-in-stock count is summed, while other fields are sourced from the SQL Server database. To achieve this, I created the Game entity in the persistence repository and mapped it to the Game model in the Domain project.